requirements.txt
selenium==4.15.2
APScheduler==3.10.4
requests==2.31.0
python-dotenv==1.0.0
